requirements.txt
## The following requirements were added by pip freeze:
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.2
pycparser==2.21
PyNaCl==1.5.0
tomli==2.0.1
yarl==1.8.1
